requirements.txt
psycopg2~=2.8.6
matplotlib~=3.3.2
speedtest-cli~=2.1.2